[Template] Rework motor library (#5)
* [Template] Rework motor library

* [Template] Add motor inversion

* [Template] Add feedforward calculator

* [Template] Added features to Bang Bang.

The bang bang calculator was missing some important features that we used on our thrower code that can tune in how the calculator should function so I added them in with getters and setters and docs.

* [Template] Removed docs by making inputs self explanatory.

I changed all of the "vel"s and "pos"s to velocity and position because it just looks nicer to me.

* [Template] Add motor neutral deadband

* [Template] Added nominal output to motor.

Nominal output allows for motors to move at their lowest possible speed. This is REALLY useful in PID controlled systems because it allows it to get right to the setpoint and not just slightly away from it. This is a better alternative to kI.

* [Template] Merge both PID calculators into one.

Both PIDCalculators were identical so I merged them into one file. We can undo this commit if the clarity of having two different types is wanted but I think this is much cleaner.

* [Template] Added multipliers to pid and ff outputs.

Added the ability to change the ratio of pid and ff outputs for finer control. Often I will balance far heavier on the PID than the ff.

* [Template] Avoid divide by 0 error if encoder ticks is set wrong.

Even if it didn't throw and error but instead returned Double.INFINITY, that woudn't do good things to the motor.

* [Template] Fix rev bug.

You typo'd

* [Template] Allow for more customizable neutral deadband settings.

Allowed for more library control of Neutral Deadband by removing the hardware configuration of 0.01 (Not very small)

* [Template] Allow for legacy TalonFXMotor creation.

To allow for a custom configured TalonFX, I included the way the legacy code worked in that it would just wrap a TalonFX instead of creating a new one.

* [Template] Added old PIDController features and legacy conversion.

* [Template] Added more filter functionality to encoders.

* [Template] Added neutral mode to Spark Max

* [Template] Swap ticks per rotation to double

This allows for the weird ticks per rotation on the johnson electric's

* [Template] Added support for custom SRX configuration.

Co-authored-by: Museum25 <91571536+mvog2501@users.noreply.github.com>